* Renamed a the particles package * Tweaked some things in ApiParticle * Updated the EventSide strings to match the new package structure And added a few todos
