+++ src/com/blazeloader/event/handlers/client/OverrideEventClientHandler.java
@@ -38,8 +38,14 @@ public interface OverrideEventClientHandler extends BLMod {
     public boolean overrideAddEntityToTracker(EntityTracker tracker, Entity entity, boolean isHandled);
 
     /**
-     * Spawns a particle into thw world.
+     * Spawns a particle into the world.
      *
+     *@deprecated TODO: This needs to be updated/removed. In 1.8 particles no longer use a string identifier. Instead there is now 
+     * EnumParticleTypes which contains most information about particles, and the EntityFX is generated by an IParticleFactory retrieved from
+     * the particlesRegistry (Map<String, IParticleFactory> EffectRenderer.field_178933_d).
+     * 
+     * A new system for handling particles has been implemented with more capabilities at ApiParticles, though it may still need some work.
+     * 
      * @param name         The name of the particle to spawn.
      * @param world        The world to spawn in.
      * @param x            The x-location to spawn at.
@@ -50,6 +56,7 @@ public interface OverrideEventClientHandler extends BLMod {
      * @param p3           Parameter 1
      * @param currParticle The particle that the previous mod generated.  Set to null if no mod has generated an particle
      * @return A generated particle, or param currParticle to disable behavior
+     * 
      */
     public EntityFX overrideSpawnParticle(String name, World world, double x, double y, double z, double p1, double p2, double p3, EntityFX currParticle);
