+++ src/com/blazeloader/event/transformers/BLEventInjectionTransformer.java
@@ -55,9 +55,9 @@ public class BLEventInjectionTransformer extends EventInjectionTransformer {
     
     protected static enum EventSide {
     	//TODO: Reminder to update these when changing package structure.
-    	CLIENT("com.blazeloader.api.client.event.EventHandlerClient"),
-    	SERVER("com.blazeloader.api.event.EventHandler"),
-    	INTERNAL("com.blazeloader.api.main.InternalEventHandler");
+    	CLIENT("com.blazeloader.event.handlers.client.EventHandlerClient"),
+    	SERVER("com.blazeloader.event.handlers.EventHandler"),
+    	INTERNAL("com.blazeloader.bl.main.InternalEventHandler");
     	
     	private final String handler;
